:recycle: refactor(ValuePicker): Remove defaultProps and propTypes.
